Make getInfo check if af_array belongs to current device.
- This behavior can be turned off optionally
