+++ src/api/c/data.cpp
@@ -539,58 +539,6 @@ af_err af_iota(af_array *result, const unsigned ndims, const dim_t * const dims,
     return AF_SUCCESS;
 }
 
-#undef INSTANTIATE
-#define INSTANTIATE(fn1, fn2)                   \
-    af_err fn1(bool *result, const af_array in) \
-    {                                           \
-        try {                                   \
-            ArrayInfo info = getInfo(in);       \
-            *result = info.fn2();               \
-        }                                       \
-        CATCHALL                                \
-            return AF_SUCCESS;                  \
-    }
-
-INSTANTIATE(af_is_empty       , isEmpty       )
-INSTANTIATE(af_is_scalar      , isScalar      )
-INSTANTIATE(af_is_row         , isRow         )
-INSTANTIATE(af_is_column      , isColumn      )
-INSTANTIATE(af_is_vector      , isVector      )
-INSTANTIATE(af_is_complex     , isComplex     )
-INSTANTIATE(af_is_real        , isReal        )
-INSTANTIATE(af_is_double      , isDouble      )
-INSTANTIATE(af_is_single      , isSingle      )
-INSTANTIATE(af_is_realfloating, isRealFloating)
-INSTANTIATE(af_is_floating    , isFloating    )
-INSTANTIATE(af_is_integer     , isInteger     )
-INSTANTIATE(af_is_bool        , isBool        )
-
-#undef INSTANTIATE
-
-af_err af_get_dims(dim_t *d0, dim_t *d1, dim_t *d2, dim_t *d3,
-                   const af_array in)
-{
-    try {
-        ArrayInfo info = getInfo(in);
-        *d0 = info.dims()[0];
-        *d1 = info.dims()[1];
-        *d2 = info.dims()[2];
-        *d3 = info.dims()[3];
-    }
-    CATCHALL
-    return AF_SUCCESS;
-}
-
-af_err af_get_numdims(unsigned *nd, const af_array in)
-{
-    try {
-        ArrayInfo info = getInfo(in);
-        *nd = info.ndims();
-    }
-    CATCHALL
-    return AF_SUCCESS;
-}
-
 template<typename T>
 static inline void eval(af_array arr)
 {
